+++ packages/mosaic/src/chart-types/base-types.ts
@@ -11,7 +11,7 @@ import type * as z from 'zod';
 import {ChartConfig, ChartType} from './chart-config';
 import {RetainedVgPlotChart} from '../VgPlotChart';
 import type {Selection} from '@uwdata/mosaic-core';
-import type {MosaicDashboardEntry} from '../dashboard/MosaicDashboardSlice';
+import type {MosaicDashboardEntry} from '../dashboard/dashboard-types';
 
 // Re-export ChartType for convenience
 export type {ChartType};
