+++ packages/mosaic/src/chart/MosaicDashboardChartRenderer.tsx
@@ -2,9 +2,9 @@ import {SpinnerPane} from '@sqlrooms/ui';
 import {BarChart3Icon} from 'lucide-react';
 import {type FC} from 'react';
 import {MosaicDashboardChartHeaderActions} from './MosaicDashboardChartHeaderActions';
+import type {ChartPanelConfig} from '../dashboard/dashboard-types';
 import {
   type MosaicDashboardPanelRenderer,
-  type ChartPanelConfig,
   type ChartPanelRendererProps,
   useStoreWithMosaicDashboard,
 } from '../dashboard/MosaicDashboardSlice';
