+++ packages/mosaic/src/dashboard/MosaicDashboardPanels.tsx
@@ -9,10 +9,10 @@ import {useMosaicDashboardContext} from './MosaicDashboardContext';
 import {MosaicDashboardEmptyState} from './MosaicDashboardEmptyState';
 import {MosaicDashboardPanelDragOverlay} from './MosaicDashboardPanelDragOverlay';
 import {MosaicDashboardPanel} from './MosaicDashboardPanel';
+import type {MosaicDashboardPanelConfig} from './dashboard-types';
 import {
   getMosaicDashboardDockId,
   MOSAIC_DASHBOARD_PANEL,
-  type MosaicDashboardPanelConfig,
   useStoreWithMosaicDashboard,
 } from './MosaicDashboardSlice';
