+++ packages/mosaic/src/dashboard/MosaicDashboardSlice.ts
@@ -27,12 +27,27 @@ import {produce} from 'immer';
 import type {ComponentType} from 'react';
 import {z} from 'zod';
 import type {ChartTypeDefinition} from '../chart-types/base-types';
+import type {ChartConfig} from '../chart-types/chart-config';
 import {type MosaicSliceState} from '../MosaicSlice';
 import {
   destroyRetainedVgPlotChart,
   type RetainedVgPlotChart,
 } from '../VgPlotChart';
-import {ChartConfig} from '../chart-types/chart-config';
+import type {
+  ChartPanelConfig,
+  MosaicDashboardEntry,
+  MosaicDashboardLayoutType,
+  MosaicDashboardPanelConfig,
+  MosaicDashboardPanelSource,
+  ProfilerPanel,
+  TextPanel,
+} from './dashboard-types';
+import {
+  MOSAIC_DASHBOARD_CHART_PANEL_TYPE,
+  MOSAIC_DASHBOARD_PROFILER_PANEL_TYPE,
+  MOSAIC_DASHBOARD_TEXT_PANEL_TYPE,
+  MosaicDashboardEntry as MosaicDashboardEntrySchema,
+} from './dashboard-types';
 
 /**
  * Panel key used for function-form panel definitions registered by
@@ -41,82 +56,6 @@ import {ChartConfig} from '../chart-types/chart-config';
  * `{ key: MOSAIC_DASHBOARD_PANEL, meta: { dashboardId, panelId } }`.
  */
 export const MOSAIC_DASHBOARD_PANEL = 'mosaic-dashboard-panel';
-export const MOSAIC_DASHBOARD_CHART_PANEL_TYPE = 'vgplot';
-export const MOSAIC_DASHBOARD_PROFILER_PANEL_TYPE = 'profiler';
-export const MOSAIC_DASHBOARD_TEXT_PANEL_TYPE = 'text';
-
-export const MosaicDashboardLayoutType = z.enum(['dock', 'grid']);
-export type MosaicDashboardLayoutType = z.infer<
-  typeof MosaicDashboardLayoutType
->;
-
-export const MosaicDashboardPanelSource = z.object({
-  tableName: z.string().optional(),
-  sqlQuery: z.string().optional(),
-});
-export type MosaicDashboardPanelSource = z.infer<
-  typeof MosaicDashboardPanelSource
->;
-
-// Profiler panel config
-export const ProfilerPanelConfig = z.object({
-  pageSize: z.number().optional(),
-});
-export type ProfilerPanelConfig = z.infer<typeof ProfilerPanelConfig>;
-
-// Text panel config
-export const TextPanelConfig = z.object({
-  content: z.string().default(''),
-  toolbarOpen: z.boolean().default(true),
-  sourcePanelOpen: z.boolean().default(false),
-});
-export type TextPanelConfig = z.infer<typeof TextPanelConfig>;
-
-// Panel configs discriminated by type
-export const ChartPanelConfig = z.object({
-  id: z.string(),
-  type: z.literal(MOSAIC_DASHBOARD_CHART_PANEL_TYPE),
-  title: z.string().default('Panel'),
-  source: MosaicDashboardPanelSource.optional(),
-  config: ChartConfig,
-});
-export type ChartPanelConfig = z.infer<typeof ChartPanelConfig>;
-
-export const ProfilerPanel = z.object({
-  id: z.string(),
-  type: z.literal(MOSAIC_DASHBOARD_PROFILER_PANEL_TYPE),
-  title: z.string().default('Panel'),
-  source: MosaicDashboardPanelSource.optional(),
-  config: ProfilerPanelConfig,
-});
-export type ProfilerPanel = z.infer<typeof ProfilerPanel>;
-
-export const TextPanel = z.object({
-  id: z.string(),
-  type: z.literal(MOSAIC_DASHBOARD_TEXT_PANEL_TYPE),
-  title: z.string().default('Text'),
-  source: MosaicDashboardPanelSource.optional(),
-  config: TextPanelConfig,
-});
-export type TextPanel = z.infer<typeof TextPanel>;
-
-// Legacy panel for backward compatibility
-export const LegacyPanelConfig = z.object({
-  id: z.string(),
-  type: z.string(),
-  title: z.string().default('Panel'),
-  source: MosaicDashboardPanelSource.optional(),
-  config: z.record(z.string(), z.unknown()).default({}),
-});
-export type LegacyPanelConfig = z.infer<typeof LegacyPanelConfig>;
-
-// Discriminated union of all panel types
-export const MosaicDashboardPanelConfig = z
-  .discriminatedUnion('type', [ChartPanelConfig, ProfilerPanel, TextPanel])
-  .or(LegacyPanelConfig);
-export type MosaicDashboardPanelConfig = z.infer<
-  typeof MosaicDashboardPanelConfig
->;
 
 export type MosaicDashboardPanelRendererProps<
   TPanel extends MosaicDashboardPanelConfig = MosaicDashboardPanelConfig,
@@ -241,19 +180,8 @@ export function createMosaicDashboardTextPanelConfig(
   };
 }
 
-export const MosaicDashboardEntry = z.object({
-  id: z.string(),
-  title: z.string().default('Dashboard'),
-  layoutType: MosaicDashboardLayoutType.default('dock'),
-  selectedTable: z.string().optional(),
-  panels: z.array(MosaicDashboardPanelConfig).default([]),
-  layout: LayoutNodeSchema.nullable().default(null),
-  updatedAt: z.number().default(0),
-});
-export type MosaicDashboardEntry = z.infer<typeof MosaicDashboardEntry>;
-
 export const MosaicDashboardSliceConfig = z.object({
-  dashboardsById: z.record(z.string(), MosaicDashboardEntry).default({}),
+  dashboardsById: z.record(z.string(), MosaicDashboardEntrySchema).default({}),
 });
 export type MosaicDashboardSliceConfig = z.infer<
   typeof MosaicDashboardSliceConfig
