+++ packages/mosaic/src/index.ts
@@ -36,6 +36,11 @@ export {
   isChartPanelConfig,
   panelHasSource,
   MOSAIC_DASHBOARD_PANEL,
+  MosaicDashboardSliceConfig,
+  resolveMosaicDashboardPanelSource,
+  useStoreWithMosaicDashboard,
+} from './dashboard/MosaicDashboardSlice';
+export {
   MOSAIC_DASHBOARD_PROFILER_PANEL_TYPE,
   MOSAIC_DASHBOARD_CHART_PANEL_TYPE,
   MOSAIC_DASHBOARD_CHART_PANEL_TYPE as MOSAIC_DASHBOARD_VGPLOT_PANEL_TYPE,
@@ -44,26 +49,25 @@ export {
   MosaicDashboardLayoutType,
   MosaicDashboardPanelConfig,
   MosaicDashboardPanelSource,
-  MosaicDashboardSliceConfig,
-  resolveMosaicDashboardPanelSource,
-  useStoreWithMosaicDashboard,
-} from './dashboard/MosaicDashboardSlice';
+} from './dashboard/dashboard-types';
 export type {
   MosaicDashboardAddPanelAction,
   MosaicDashboardAddPanelActionContext,
   CreateMosaicDashboardSliceProps,
-  MosaicDashboardEntry as MosaicDashboardEntryType,
-  MosaicDashboardPanelConfig as MosaicDashboardPanelConfigType,
   MosaicDashboardPanelRenderer,
   MosaicDashboardPanelRendererProps,
-  MosaicDashboardPanelSource as MosaicDashboardPanelSourceType,
   MosaicDashboardSliceConfig as MosaicDashboardSliceConfigType,
   MosaicDashboardSliceState,
   MosaicDashboardStoreState,
+} from './dashboard/MosaicDashboardSlice';
+export type {
+  MosaicDashboardEntry as MosaicDashboardEntryType,
+  MosaicDashboardPanelConfig as MosaicDashboardPanelConfigType,
+  MosaicDashboardPanelSource as MosaicDashboardPanelSourceType,
   ChartPanelConfig,
   TextPanel,
   TextPanelConfig,
-} from './dashboard/MosaicDashboardSlice';
+} from './dashboard/dashboard-types';
 export {
   createMosaicColorLegendPlot,
   MosaicColorLegend,
